+++ R/response.R
@@ -25,7 +25,7 @@
 #' @name responses
 #'
 #' @export
-response <- function(body, headers = content_html(), status = 200L){
+response <- function(body, headers = NULL, status = 200L){
   assert_that(not_missing(body))
   res <- list(status = as.integer(status), headers = headers, body = convert_body(body))
   construct_response(res)
@@ -33,14 +33,14 @@ response <- function(body, headers = content_html(), status = 200L){
 
 #' @rdname responses
 #' @export
-response_404 <- function(body = "404: Not found", headers = content_html(), status = 404L){
+response_404 <- function(body = "404: Not found", headers = NULL, status = 404L){
   res <- list(status = as.integer(status), headers = headers, body = convert_body(body))
   construct_response(res)
 }
 
 #' @rdname responses
 #' @export
-response_500 <- function(body = "500: Server Error", headers = content_html(), status = 500L){
+response_500 <- function(body = "500: Server Error", headers = NULL, status = 500L){
   res <- list(status = as.integer(status), headers = headers, body = convert_body(body))
   construct_response(res)
 }
@@ -49,13 +49,14 @@ convert_body <- function(body) {
   if(inherits(body, "AsIs"))
     return(body)
 
-  if(is.character(body))
+  if(is.character(body) && length(body) == 1L)
     return(body)
 
   if(is.factor(body) || inherits(body, "shiny.tag"))
     return(as.character(body))
 
-  return(body)
+  as.character(body) |> 
+    paste0(collapse = "")
 }
 
 #' Construct Response
@@ -99,7 +100,7 @@ Response <- R6::R6Class(
 #' @param body Body of the response.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response, if `NULL` uses `self$status`.
-    send = function(body, headers = content_html(), status = NULL){
+    send = function(body, headers = NULL, status = NULL){
       headers <- private$.get_headers(headers)
       response(status = private$.get_status(status), headers = headers, body = as.character(body))
     },
@@ -108,7 +109,7 @@ Response <- R6::R6Class(
 #' @param ... Passed to `...` of `sprintf`.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response, if `NULL` uses `self$status`.
-    sendf = function(body, ..., headers = content_html(), status = NULL){
+    sendf = function(body, ..., headers = NULL, status = NULL){
       body <- sprintf(body, ...)
       headers <- private$.get_headers(headers)
       response(status = private$.get_status(status), headers = headers, body = as.character(body))
@@ -117,7 +118,7 @@ Response <- R6::R6Class(
 #' @param body Body of the response.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response, if `NULL` uses `self$status`.
-    text = function(body, headers = content_plain(), status = NULL){
+    text = function(body, headers = NULL, status = NULL){
       headers <- private$.get_headers(headers)
       response(status = private$.get_status(status), headers = headers, body = as.character(body))
     },
@@ -125,7 +126,7 @@ Response <- R6::R6Class(
 #' @param file File to send.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response.
-    send_file = function(file, headers = content_html(), status = NULL){
+    send_file = function(file, headers = NULL, status = NULL){
       assert_that(not_missing(file))
       self$render(file, data = list(), status = private$.get_status(status), headers = headers)
     },
@@ -145,7 +146,7 @@ Response <- R6::R6Class(
 #' @param data List to fill `[% tags %]`.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response, if `NULL` uses `self$status`.
-    render = function(file, data = list(), headers = content_html(), status = NULL){
+    render = function(file, data = list(), headers = NULL, status = NULL){
       assert_that(not_missing(file))
       assert_that(has_file(file))
 
@@ -224,7 +225,7 @@ Response <- R6::R6Class(
 #' @param data List to fill `[% tags %]`.
 #' @param headers HTTP headers to set.
 #' @param status Status of the response, if `NULL` uses `self$status`.
-    md = function(file, data = list(), headers = content_html(), status = NULL) {
+    md = function(file, data = list(), headers = NULL, status = NULL) {
       check_installed("commonmark")
       self$render(file, data, headers, status)
     },
@@ -538,9 +539,18 @@ Response <- R6::R6Class(
 
       status
     },
-    .get_headers = function(headers = list()){
+    .get_headers = function(headers = NULL){
       private$.render_cookies()
-      modifyList(private$.headers, headers)
+      print(headers)
+      heads <- private$.headers
+
+      if(!is.null(headers))
+        heads <- modifyList(heads, headers)
+
+      if(is.null(heads[["Content-Type"]]))
+        heads <- modifyList(heads, content_html())
+
+      return(heads)
     },
     .run_post_hooks = function(file_content, ext) {
       if(length(private$.postHooks) == 0) {